+++ tools/ColorTool/ColorTool/ColorScheme.cs
@@ -1,9 +1,10 @@
 ﻿//
-//    Copyright (C) Microsoft.  All rights reserved.
+// Copyright (C) Microsoft.  All rights reserved.
 // Licensed under the terms described in the LICENSE file in the root of this project.
 //
 
 using System;
+using System.Drawing;
 using System.Linq;
 
 namespace ColorTool
@@ -13,6 +14,15 @@ namespace ColorTool
         public uint[] colorTable = null;
         public ConsoleAttributes consoleAttributes;
 
+        public Color this[int index] => UIntToColor(colorTable[index]);
+
+        private static Color UIntToColor(uint color)
+        {
+            byte r = (byte)(color >> 0);
+            byte g = (byte)(color >> 8);
+            byte b = (byte)(color >> 16);
+            return Color.FromArgb(r, g, b);
+        }
 
         public int CalculateIndex(uint value) =>
             colorTable.Select((color, idx) => Tuple.Create(color, idx))
