+++ tools/ColorTool/ColorTool/ConsoleAPI.cs
@@ -1,5 +1,5 @@
 ﻿//
-//    Copyright (C) Microsoft.  All rights reserved.
+// Copyright (C) Microsoft.  All rights reserved.
 // Licensed under the terms described in the LICENSE file in the root of this project.
 //
 
@@ -59,6 +59,9 @@ namespace ColorTool
         }
 
         public static int STD_OUTPUT_HANDLE = -11;
+
+
+        public static IntPtr GetStdOutputHandle() => GetStdHandle(STD_OUTPUT_HANDLE);
 
         [DllImport("kernel32.dll", SetLastError = true)]
         public static extern IntPtr GetStdHandle(int nStdHandle);
