+++ tools/ColorTool/ColorTool/ConsoleAttributes.cs
@@ -1,5 +1,5 @@
 ﻿//
-//    Copyright (C) Microsoft.  All rights reserved.
+// Copyright (C) Microsoft.  All rights reserved.
 // Licensed under the terms described in the LICENSE file in the root of this project.
 //
 
@@ -12,6 +12,5 @@ namespace ColorTool
 
         public uint? popupForeground;
         public uint? popupBackground;
-
     }
 }
