+++ tools/ColorTool/ColorTool/SchemeParsers/ISchemeParser.cs
@@ -1,9 +1,9 @@
 ﻿//
-//    Copyright (C) Microsoft.  All rights reserved.
+// Copyright (C) Microsoft.  All rights reserved.
 // Licensed under the terms described in the LICENSE file in the root of this project.
 //
 
-namespace ColorTool
+namespace ColorTool.SchemeParsers
 {
     interface ISchemeParser
     {
